+++ internal/job/checkout_fetch.go
@@ -10,9 +10,7 @@ import (
 	"github.com/buildkite/agent/v3/internal/shell"
 )
 
-// refspecKind is the category of git refspec a fetch targets. It is classified
-// once in fetchSource and then drives both the git.refspec_kind span attribute
-// and the fetch dispatch, so the label and the fetch behaviour cannot drift apart.
+// refspecKind is the category of git refspec a fetch targets.
 type refspecKind string
 
 const (
@@ -38,21 +36,12 @@ const (
 // flags — the caller decides based on sparse-checkout state and user-supplied
 // filters.
 func (e *Executor) fetchSource(ctx context.Context, addBloblessFilter bool) (retErr error) {
-	// The git.fetch span is started here (inside the function that owns the
-	// operation) so its attributes — decided below — are set directly on the
-	// in-scope span, avoiding any signature change to fetchSource.
-	//
-	// Note: this span covers the whole fetch, including the retry loops inside
-	// gitFetch / gitFetchWithFallback (up to 10 attempts, ~2m). A long span may
-	// therefore be several retried fetches rather than one slow fetch; it does
-	// not break down per attempt.
+	// Start span here so attributes can be set on the in-scope span; covers the
+	// whole fetch including retries (up to 10 attempts, ~2m), not per-attempt.
 	span, ctx := e.traceGitOpSpan(ctx, "git.fetch")
 	defer func() { span.FinishWithError(retErr) }()
 
-	// Classify the refspec kind once, up front. The work switch below dispatches
-	// on this same value rather than re-evaluating the conditions, so the label
-	// and the fetch behaviour stay in lockstep. Done before the skip-fetch
-	// short-circuit so a skipped fetch is still labelled accurately.
+	// Classify the refspec kind once and dispatch on it in the switch below.
 	kind := refspecCommit
 	switch {
 	case e.RefSpec != "":
@@ -66,6 +55,7 @@ func (e *Executor) fetchSource(ctx context.Context, addBloblessFilter bool) (ret
 	case e.Commit == "HEAD":
 		kind = refspecBranch
 	}
+
 	span.AddAttributes(map[string]string{
 		"git.pull_request": strconv.FormatBool(e.PullRequest != "false"),
 		"git.refspec_kind": string(kind),
@@ -76,7 +66,9 @@ func (e *Executor) fetchSource(ctx context.Context, addBloblessFilter bool) (ret
 	// as the commit objects are already reachable and fetching is redundant.
 	skipFetch := e.GitSkipFetchExistingCommits && e.Commit != "HEAD" &&
 		hasGitCommit(ctx, e.shell, ".git", e.Commit)
+
 	span.AddAttributes(map[string]string{"git.skipped": strconv.FormatBool(skipFetch)})
+
 	if skipFetch {
 		e.shell.Commentf("Commit %q already exists locally, skipping fetch", e.Commit)
 		return nil
