+++ internal/job/tracing.go
@@ -341,6 +341,26 @@ func toOpenTelemetryAttributes(extras map[string]any) ([]attribute.KeyValue, map
 	return attrs, unknownAttrTypes
 }
 
+// traceGitOp runs fn in a child span named `name`, finishing it with fn's error.
+// No-op if --trace-git-checkout is off.
+func (e *Executor) traceGitOp(ctx context.Context, name string, fn func(context.Context) error) (err error) {
+	if !e.TraceGitCheckout {
+		return fn(ctx)
+	}
+	span, ctx := tracetools.StartSpanFromContext(ctx, name, e.TracingBackend)
+	defer func() { span.FinishWithError(err) }()
+	return fn(ctx)
+}
+
+// traceGitOpSpan is like traceGitOp but returns the span so the caller can set
+// attributes and must call FinishWithError. Returns a NoopSpan when tracing is off.
+func (e *Executor) traceGitOpSpan(ctx context.Context, name string) (tracetools.Span, context.Context) {
+	if !e.TraceGitCheckout {
+		return &tracetools.NoopSpan{}, ctx
+	}
+	return tracetools.StartSpanFromContext(ctx, name, e.TracingBackend)
+}
+
 func (e *Executor) implementationSpecificSpanName(otelName, ddName string) string {
 	switch e.TracingBackend {
 	case tracetools.BackendDatadog:
